returned interval type

+++ Arduino/LEDController.cpp
@@ -21,7 +21,7 @@ void LEDController::changeState(uint8_t new_state)
     return;
 }
 
-bool LEDController::onIntervalPassed(unsigned int interval)
+bool LEDController::onIntervalPassed(unsigned long interval)
 {
     unsigned long curr_time = millis();
     if (curr_time - _prev_time >= interval)
@@ -32,7 +32,7 @@ bool LEDController::onIntervalPassed(unsigned int interval)
     return false;
 }
 
-bool LEDController::offIntervalPassed(unsigned int interval)
+bool LEDController::offIntervalPassed(unsigned long interval)
 {
     unsigned long curr_time = millis();
     if (curr_time - _prev_time >= interval)
